Move everything out of genes.py and in their place in separate files
