+++ src/roxene/genes/__init__.py
@@ -1,0 +1,7 @@
+from .create_neuron import CreateNeuron
+from .connect_neurons import ConnectNeurons
+from .create_input_cell import CreateInputCell
+from .rotate_cells import RotateCells
+from .composite_gene import CompositeGene
+
+__all__ = ['CreateNeuron', 'ConnectNeurons', 'RotateCells', 'CreateInputCell', 'CompositeGene']
