+++ src/roxene/mutagens/__init__.py
@@ -1,6 +1,6 @@
 from .mutagen import Mutagen
-from .create_neuron_mutagen import CreateNeuronMutagen
+from .create_neuron_mutagen import CreateNeuronMutagen, CNLayer
 from .composite_gene_split_mutagen import CompositeGeneSplitMutagen
 from .util import wiggle
 
-__all__ = ['Mutagen', 'CreateNeuronMutagen', 'CompositeGeneSplitMutagen', 'wiggle']
+__all__ = ['Mutagen', 'CreateNeuronMutagen', 'CompositeGeneSplitMutagen', 'wiggle', 'CNLayer']
