+++ src/roxene/tic_tac_toe/environment.py
@@ -12,8 +12,8 @@ from .outcome import Outcome
 from .players import REQUIRED_INPUTS, REQUIRED_OUTPUTS, Player
 from .population import Population
 from .trial import Trial
-from ..genes import CompositeGene, CreateNeuron, ConnectNeurons, RotateCells, CNLayer
-from ..mutagens import CreateNeuronMutagen, Mutagen, wiggle
+from ..genes import CompositeGene, CreateNeuron, ConnectNeurons, RotateCells
+from ..mutagens import CreateNeuronMutagen, Mutagen, wiggle, CNLayer
 from ..organism import Organism, Gene
 from ..util import random_neuron_state
